+++ pkg/alertmanager/config.go
@@ -44,6 +44,16 @@ func createUsableGrafanaConfig(logger log.Logger, gCfg alertspb.GrafanaAlertConf
 		amCfg.AlertmanagerConfig.Global = cfg.Global
 	}
 
+	// If configured, use the SMTP From address sent by Grafana.
+	// TODO: Remove once it's sent in the SmtpConfig field.
+	if gCfg.SmtpFrom != "" {
+		if amCfg.AlertmanagerConfig.Global == nil {
+			defaultGlobals := config.DefaultGlobalConfig()
+			amCfg.AlertmanagerConfig.Global = &defaultGlobals
+		}
+		amCfg.AlertmanagerConfig.Global.SMTPFrom = gCfg.SmtpFrom
+	}
+
 	// We want to:
 	// 1. Remove duplicate receivers and keep the last receiver occurrence if there are conflicts. This is based on the upstream implementation.
 	// 2. Maintain a consistent ordering and preferably original ordering. Otherwise, change detection will be impacted.
@@ -90,6 +100,9 @@ func createUsableGrafanaConfig(logger log.Logger, gCfg alertspb.GrafanaAlertConf
 		KeyFile:      g.HTTPConfig.TLSConfig.KeyFile,
 		SkipVerify:   !g.SMTPRequireTLS,
 		SentBy:       fmt.Sprintf("Mimir v%s", version.Version),
+
+		// TODO: Remove once it's sent in SmtpConfig.
+		StaticHeaders: gCfg.StaticHeaders,
 	}
 
 	// Patch the base config with the custom SMTP config sent by Grafana.