+++ pkg/streamingpromql/operators/aggregations/count.go
@@ -87,6 +87,5 @@ func (g *CountGroupAggregationGroup) ComputeOutputSeries(_ types.ScalarData, tim
 }
 
 func (g *CountGroupAggregationGroup) Close(memoryConsumptionTracker *limiter.MemoryConsumptionTracker) {
-	types.Float64SlicePool.Put(g.values, memoryConsumptionTracker)
-	g.values = nil
+	types.Float64SlicePool.Put(&g.values, memoryConsumptionTracker)
 }