+++ pkg/streamingpromql/operators/functions/times.go
@@ -45,8 +45,7 @@ func timeWrapperFunc(f func(t time.Time) float64) InstantVectorSeriesFunction {
 	return func(seriesData types.InstantVectorSeriesData, _ []types.ScalarData, _ types.QueryTimeRange, memoryConsumptionTracker *limiter.MemoryConsumptionTracker) (types.InstantVectorSeriesData, error) {
 
 		// we don't do time based function on histograms
-		types.HPointSlicePool.Put(seriesData.Histograms, memoryConsumptionTracker)
-		seriesData.Histograms = nil
+		types.HPointSlicePool.Put(&seriesData.Histograms, memoryConsumptionTracker)
 
 		for i := range seriesData.Floats {
 			t := time.Unix(int64(seriesData.Floats[i].F), 0).UTC()